Setting amount fields to allow doubles

+++ backend/Dtos/SavingGoal/SavingGoalDto.cs
@@ -6,7 +6,7 @@ namespace backend.Dtos.SavingGoal
     {
         public int Id { get; set; }
         public string Details { get; set; } = String.Empty;
-        public int Goal { get; set; }
+        public double Goal { get; set; }
         public int? FixedContribution { get; set; }
         public int? Periodicity { get; set; }
         public List<Models.GoalContribution> GoalContributions { get; set; } = new List<Models.GoalContribution>();
